+++ hw05-aop/src/main/java/org/kt/proxy/Ioc.java
@@ -1,7 +1,7 @@
 package org.kt.proxy;
 
 import org.kt.share.TestLogging;
-import org.kt.share.TestLoggingInterface;
+import org.kt.share.TestLoggingImpl;
 import org.kt.share.annotations.Log;
 
 import java.lang.annotation.Annotation;
@@ -15,21 +15,21 @@ public class Ioc {
     private Ioc() {
     }
 
-    static TestLoggingInterface createTestLogging() {
-        InvocationHandler handler = new TestLoggingInvocationHandler(new TestLogging());
-        return (TestLoggingInterface) Proxy.newProxyInstance(Ioc.class.getClassLoader(),
-                new Class<?>[]{TestLoggingInterface.class}, handler);
+    static TestLogging createTestLogging() {
+        InvocationHandler handler = new Handler(new TestLoggingImpl());
+        return (TestLogging) Proxy.newProxyInstance(Ioc.class.getClassLoader(),
+                new Class<?>[]{TestLogging.class}, handler);
     }
 
-    static class TestLoggingInvocationHandler implements InvocationHandler {
-        private final TestLoggingInterface myClass;
+    static class Handler implements InvocationHandler {
+        private final Object origin;
         final private Map<String, Method> methodLogs = new HashMap<>();
         private String getMethodSignature(Method method) {
             return method.getName() + Arrays.toString(method.getParameterTypes());
         }
-        TestLoggingInvocationHandler(TestLoggingInterface myClass) {
-            this.myClass = myClass;
-            for (Method method : myClass.getClass().getDeclaredMethods()) {
+        Handler(Object origin) {
+            this.origin = origin;
+            for (Method method : origin.getClass().getDeclaredMethods()) {
                 Annotation[] annotations = method.getAnnotations();
                 for (Annotation annotation : annotations) {
                     if (annotation instanceof Log) {
@@ -39,20 +39,19 @@ public class Ioc {
             }
         }
 
-
         @Override
         public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
             if (methodLogs.containsKey(getMethodSignature(method))) {
                 System.out.println("execution method: " + method.getName() + ", params: " + (args != null ? Arrays.toString(args) : ""));
             }
 
-            return method.invoke(myClass, args);
+            return method.invoke(origin, args);
         }
 
         @Override
         public String toString() {
-            return "TestLoggingInvocationHandler{" +
-                    "testLoggingClass=" + myClass +
+            return "Handler{" +
+                    "myClass=" + origin +
                     '}';
         }
     }
