+++ hw05-aop/src/main/java/org/kt/share/TestLogging.java
@@ -1,29 +1,12 @@
 package org.kt.share;
 
-import org.kt.share.annotations.Log;
-
-public class TestLogging implements TestLoggingInterface {
-
-    @Log
-    public void calculation(int param1) {
-        System.out.println("Hello world!");
-    }
-
-    public void calculation(int param1, int param2) {}
-
-    @Log
-    public void calculation(int param1, int param2, String param3) {}
-
-    @Log
-    public void method1() {}
-
-    public void method1(int param1) {}
-
-    @Log
-    public void method1(String param1) {}
-
-    public void method2(int param1, int param2) {}
-
-    @Log
-    public void method2(int param1, int param2, String param3) {}
+public interface TestLogging {
+    void calculation(int param1);
+    void calculation(int param1, int param2);
+    void calculation(int param1, int param2, String param3);
+    void method1();
+    void method1(int param1);
+    void method1(String param1);
+    void method2(int param1, int param2);
+    void method2(int param1, int param2, String param3);
 }
