--- hw05-aop/src/main/java/org/kt/share/TestLoggingInterface.java
@@ -1,12 +1,0 @@
-package org.kt.share;
-
-public interface TestLoggingInterface {
-    void calculation(int param1);
-    void calculation(int param1, int param2);
-    void calculation(int param1, int param2, String param3);
-    void method1();
-    void method1(int param1);
-    void method1(String param1);
-    void method2(int param1, int param2);
-    void method2(int param1, int param2, String param3);
-}
